moving debugger to dataset and switching over all references to the new module
